% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbinom.R
\name{cbinom}
\alias{cbinom}
\title{Binomial Distribution of N Bernouli TRials}
\usage{
cbinom(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Choose of how many times to run}

\item{n}{Number of Bernoulli Trials}

\item{p}{Probability of Success}
}
\value{
The equivalent of dbinom function
}
\description{
Binomial Distribution of N Bernouli TRials
}
\examples{
\dontrun{cbinom(iter=1000, n=5,p=.5)}
}
