% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{modeit}
\alias{modeit}
\title{A way to get mode}
\usage{
modeit(v)
}
\arguments{
\item{v}{A quantitative varaible}
}
\value{
The mode for that set of data
}
\description{
A way to get mode
}
\examples{
\dontrun{modeit(v=data)}
}
