% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycboot.R
\name{mycboot}
\alias{mycboot}
\title{Boot Strap Function}
\usage{
mycboot(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Iterations}

\item{x}{Sample Data}

\item{fun}{Function=mean,var,sd,ect.}

\item{alpha}{Confidence}

\item{cx}{Text Zoom}

\item{...}{Add Extras}
}
\value{
Returns a confidence interval for chosen function
}
\description{
Boot Strap Function
}
\examples{
mycboot(iter=10000,x=runif(20,min=0,max=5),fun=mean,alpha=0.05)
}
