% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltunif.R
\name{mycltunif}
\alias{mycltunif}
\title{Central Limit Theory Function for Runif}
\usage{
mycltunif(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{Sample Size}

\item{iter}{How many Iterations}

\item{a}{Left Side of Unif}

\item{b}{Right Side of Unif}
}
\value{
A random sample from runif that gets packaged for mean distribution to mimic a normal distribution
}
\description{
Central Limit Theory Function for Runif
}
\examples{
mycltunif(10,10000,0,10)
}
