% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlikfun.R
\name{mymaxlikfun}
\alias{mymaxlikfun}
\title{Maximum Likelihood Function}
\usage{
mymaxlikfun(lfun, x, param, ...)
}
\arguments{
\item{lfun}{Functions =logbin,logpoiss, and logexp}

\item{x}{Sample Data}

\item{param}{Param=seq(0,1,length=10000) binom or seq(0,20,length=10000)}

\item{...}{Add extras}
}
\value{
Returns point estimate of choice
}
\description{
Maximum Likelihood Function
}
\examples{
\dontrun{mymaxlikfun(x=c(1,2,3,4,5,6,7),param=seq(0,1,length=10000),lfun=logbin,xlab=expression(pi),main="Binomial",cex.main=2)}
}
