% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Dnorm Function for Shading Area under the Curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Input mean}

\item{sigma}{Input sd}

\item{a}{Input P(Y<=a)}
}
\value{
Same as dnorm, but with a shading function
}
\description{
Dnorm Function for Shading Area under the Curve
}
\examples{
myncurve(mu=5,sigma=10,a=6)
}
