% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Title}
\usage{
ntickets(N = 200, g = 0.02, p = 0.95)
}
\arguments{
\item{N}{Input # of Tickets Sold}

\item{g}{Input Prob of Overbooking}

\item{p}{Input Prob}
}
\value{
Creates n Tickets allowable to sell to maximize profits
}
\description{
Title
}
\examples{
ntickets(200,.02,.95)
}
